rename utils package to resourcepoll
