+++ pkg/controller/servicebindingrequest/planner/planner.go
@@ -3,7 +3,7 @@ package planner
 import (
 	"context"
 	"fmt"
-	"github.com/redhat-developer/service-binding-operator/pkg/utils"
+	"github.com/redhat-developer/service-binding-operator/pkg/resourcepoll"
 	"strings"
 
 	"github.com/go-logr/logr"
@@ -101,7 +101,7 @@ func (p *Planner) searchCR(kind string) (*ustrv1.Unstructured, error) {
 	}}
 	namespacedName := types.NamespacedName{Namespace: p.ns, Name: resourceRef}
 
-	err = utils.WaitUntilResourceFound(p.client, namespacedName, &cr)
+	err = resourcepoll.WaitUntilResourceFound(p.client, namespacedName, &cr)
 	if err != nil {
 		return nil, err
 	}
