+++ pkg/resourcepoll/wait.go
@@ -1,4 +1,4 @@
-package utils
+package resourcepoll
 
 import (
 	"context"
